// src/api/openapi/models/RouteItemResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { StopItemResponseDto } from './StopItemResponseDto';
import {
    StopItemResponseDtoFromJSON,
    StopItemResponseDtoFromJSONTyped,
    StopItemResponseDtoToJSON,
    StopItemResponseDtoToJSONTyped,
} from './StopItemResponseDto';

/**
 * 
 * @export
 * @interface RouteItemResponseDto
 */
export interface RouteItemResponseDto {
    /**
     * 
     * @type {Date}
     * @memberof RouteItemResponseDto
     */
    start: Date;
    /**
     * 
     * @type {Date}
     * @memberof RouteItemResponseDto
     */
    end: Date;
    /**
     * 
     * @type {StopItemResponseDto}
     * @memberof RouteItemResponseDto
     */
    from: StopItemResponseDto;
    /**
     * 
     * @type {StopItemResponseDto}
     * @memberof RouteItemResponseDto
     */
    to: StopItemResponseDto;
}

/**
 * Check if a given object implements the RouteItemResponseDto interface.
 */
export function instanceOfRouteItemResponseDto(value: object): value is RouteItemResponseDto {
    if (!('start' in value) || value['start'] === undefined) return false;
    if (!('end' in value) || value['end'] === undefined) return false;
    if (!('from' in value) || value['from'] === undefined) return false;
    if (!('to' in value) || value['to'] === undefined) return false;
    return true;
}

export function RouteItemResponseDtoFromJSON(json: any): RouteItemResponseDto {
    return RouteItemResponseDtoFromJSONTyped(json, false);
}

export function RouteItemResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): RouteItemResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'start': (new Date(json['start'])),
        'end': (new Date(json['end'])),
        'from': StopItemResponseDtoFromJSON(json['from']),
        'to': StopItemResponseDtoFromJSON(json['to']),
    };
}

export function RouteItemResponseDtoToJSON(json: any): RouteItemResponseDto {
    return RouteItemResponseDtoToJSONTyped(json, false);
}

export function RouteItemResponseDtoToJSONTyped(value?: RouteItemResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'start': value['start'].toISOString(),
        'end': value['end'].toISOString(),
        'from': StopItemResponseDtoToJSON(value['from']),
        'to': StopItemResponseDtoToJSON(value['to']),
    };
}