// src/api/openapi/models/StopItemResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PlaceItemResponseDto } from './PlaceItemResponseDto';
import {
    PlaceItemResponseDtoFromJSON,
    PlaceItemResponseDtoFromJSONTyped,
    PlaceItemResponseDtoToJSON,
    PlaceItemResponseDtoToJSONTyped,
} from './PlaceItemResponseDto';

/**
 * 
 * @export
 * @interface StopItemResponseDto
 */
export interface StopItemResponseDto {
    /**
     * 
     * @type {PlaceItemResponseDto}
     * @memberof StopItemResponseDto
     */
    place: PlaceItemResponseDto;
}

/**
 * Check if a given object implements the StopItemResponseDto interface.
 */
export function instanceOfStopItemResponseDto(value: object): value is StopItemResponseDto {
    if (!('place' in value) || value['place'] === undefined) return false;
    return true;
}

export function StopItemResponseDtoFromJSON(json: any): StopItemResponseDto {
    return StopItemResponseDtoFromJSONTyped(json, false);
}

export function StopItemResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StopItemResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'place': PlaceItemResponseDtoFromJSON(json['place']),
    };
}

export function StopItemResponseDtoToJSON(json: any): StopItemResponseDto {
    return StopItemResponseDtoToJSONTyped(json, false);
}

export function StopItemResponseDtoToJSONTyped(value?: StopItemResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'place': PlaceItemResponseDtoToJSON(value['place']),
    };
}